Adjust .csproj file
Updated the .csproj file. To compile properly, this file will need to be located elsewhere. See the README.md file for further details.

The previous setup depended a lot on packages that may not even be available, as well as referencing file paths that were fine for the original creator, but were dated and potentially difficult for the user to ensure existed.