The Big Rebrand
Short version:
Project has been renamed to better represent future ambitions.
Long version:
Considering a lot of what I've worked on for the past few months, especially the features yet to be integrated, I wasn't sure TNH Tweaker was a great fit for the name of this project. I fear about it in it's current state, I don't want to take Devyn's hard work and rebrand it as my own, but for multiple reasons including compatibility, it's likely better I rebrand this before it goes public, and I don't have to worry about big future compatibility issues.
Devyn's payment link and whatnot will be preserved in the README until I've created enough to feel worthy of it.